requirements.txt
Flask~=2.3.2
stripe~=5.4.0
kafka-python~=2.0.2
requests~=2.31.0
python-dotenv